+++ src/components/profile/index.ts
@@ -4,4 +4,5 @@ export { ProfileHeader } from './ProfileHeader';
 export { ProfileStats } from './ProfileStats';
 export { ProfileTabs } from './ProfileTabs';
 export { UnclaimedBanner } from './UnclaimedBanner';
+export { ClaimModal } from './ClaimModal';
 export { VideoGrid } from './VideoGrid';